Keep html,body on height:100%; use body for scroll event
`html, body { height: 100%; }` was overridden in `slidebars.min.css`,
so it was impossible to create content elements with 100% page height.

Changing that required to also change `antimatter.js` to bind the scroll
event to the `body` element instead of the document because Slidebars'
`x-overflow: hidden` (see http://stackoverflow.com/questions/17665080).